requirements.txt
psycopg2-binary
pandas==1.0.4
requests